Add feature to exclude input values close to zero
Implemented feature as a configuration value that can be set as the
number of places that can be zero before the value is considered zero.
This includes the decimal place before the comma.
The implementation is based on setting an invalid Polar coordinate
using std::nan so that the Drawer can recognise that it should not be
drawn. For every IOVector, there are now to more checks.